made random agents seedable for reproducable games.
