+++ src/agent/mod.rs
@@ -1,10 +1,9 @@
-
 use crate::game::scoped_state::ScopedGameState;
-use crate::utils::Action;
+use crate::action::Action;
 
 pub mod random;
 pub mod human;
 
-pub trait Agent {
+pub trait Agent{
     fn decide_action(&mut self, state: &ScopedGameState) -> Action;
 }
