+++ src/game/mod.rs
@@ -1,11 +1,12 @@
-use rand::{Rng, SeedableRng};
+use rand::Rng;
 use rand_chacha::ChaCha8Rng;
 
 use crate::player::Player;
 use crate::card::{Card, Suit};
-use crate::utils::{Action, FlippedChoice};
+use crate::action::{Action, FlippedChoice};
 use crate::dealer::Dealer;
 use crate::judger::Judger;
+use crate::utils::get_rdm_gen;
 
 pub mod scoped_state;
 use self::scoped_state::ScopedGameState;
@@ -38,7 +39,7 @@ impl EuchreGame {
             seed: Option<u64>,
         ) -> EuchreGame {
 
-        let mut gen: ChaCha8Rng = Self::get_rdm_gen(seed);
+        let mut gen: ChaCha8Rng = get_rdm_gen(seed);
 
         let deal_id = determine_dealer(dealer_id, &mut gen);
         let curr_p_id = (deal_id + 1) % 4;
@@ -77,14 +78,6 @@ impl EuchreGame {
 
     pub fn is_over(&self) -> bool {
         self.is_over
-    }
-
-
-    fn get_rdm_gen(seed: Option<u64>) -> ChaCha8Rng {
-        match seed {
-            Some(num) => {ChaCha8Rng::seed_from_u64(num)},
-            None => {ChaCha8Rng::from_entropy()}
-        }
     }
 
     /// Get the current game state as the current player sees it.
